# docs/swagger.yaml
basePath: /service/api
definitions:
...
  dto.BaseResponse:
    properties:
      data: {}
      message_en:
        example: Success
        type: string
      message_th:
        example: สำเร็จ
        type: string
      status:
        example: success
        type: string
      status_code:
        example: 200
        type: integer
    type: object
...
  dto.CreateIncomeDTO:
    properties:
      amount:
        type: number
      bank_id:
        type: string
      currency:
        type: string
      description:
        type: string
      note:
        type: string
      payment_method:
        type: string
      reference_no:
        type: string
      transaction_category_id:
        type: string
      txn_date:
        type: string
    required:
    - amount
    - bank_id
    - currency
    - description
    - transaction_category_id
    - txn_date
    type: object
...
  dto.IncomeSummaryDTO:
    properties:
      total_all:
        type: number
      total_this_month:
        type: number
      total_today:
        type: number
    type: object
...
  dto.UpdateIncomeDTO:
    properties:
      amount:
        type: number
      bank_id:
        type: string
      currency:
        type: string
      description:
        type: string
      note:
        type: string
      payment_method:
        type: string
      reference_no:
        type: string
      transaction_category_id:
        type: string
      txn_date:
        type: string
    required:
    - amount
    - bank_id
    - currency
    - description
    - transaction_category_id
    - txn_date
    type: object
...
host: api.dev.rkp-media.com
info:
  contact: {}
  description: This is an ERP API demo.
  title: ERP Demo API
  version: "1.0"
paths:
...
  /v1/in-come/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an income by its ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Delete Income by ID
      tags:
      - Income
    get:
      consumes:
      - application/json
      description: Get a single income by its ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get Income by ID
      tags:
      - Income
    put:
      consumes:
      - application/json
      description: Update an existing income by its ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Income
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIncomeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Update Income by ID
      tags:
      - Income
  /v1/in-come/create:
    post:
      consumes:
      - application/json
      description: Create a new income
      parameters:
      - description: Create Income
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIncomeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Create In Come
      tags:
      - Income
  /v1/in-come/list:
    get:
      consumes:
      - application/json
      description: List all incomes with pagination and optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort by field
        enum:
        - created_at
        - updated_at
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List In Comes
      tags:
      - Income
  /v1/in-come/summary:
    get:
      consumes:
      - application/json
      description: Get summary of income for today, this month, and all time
      parameters:
      - description: Bank ID to filter incomes
        in: query
        name: bank_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.IncomeSummaryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Income Summary by Filter
      tags:
      - Income
...
swagger: "2.0"